Rename to Asn1BerCodec
